Refactor dimension index sequence
