+++ src/highdicom/__init__.py
@@ -61,6 +61,7 @@ from highdicom.enum import (
 )
 from highdicom import frame
 from highdicom.image import (
+    DimensionIndexSequence,
     Image,
     imread,
     get_volume_from_series,
@@ -89,6 +90,7 @@ __all__ = [
     'ContentQualificationValues',
     'ContributingEquipment',
     'CoordinateSystemNames',
+    'DimensionIndexSequence',
     'DimensionOrganizationTypeValues',
     'Image',
     'InterpolationMethods',
