+++ src/highdicom/_module_utils.py
@@ -1,5 +1,5 @@
 from enum import Enum
-from typing import Any, Dict, List, Optional, Union
+from typing import Any
 from collections.abc import Sequence
 
 from pydicom import Dataset
