+++ src/highdicom/pm/content.py
@@ -1,19 +1,15 @@
 from collections.abc import Sequence
+import warnings
 
 import numpy as np
-from pydicom.datadict import keyword_for_tag, tag_for_keyword
 from pydicom.dataset import Dataset
-from pydicom.sequence import Sequence as DataElementSequence
 from pydicom.sr.coding import Code
-from highdicom._module_utils import is_multiframe_image
 
-from highdicom.content import PlanePositionSequence
 from highdicom.enum import CoordinateSystemNames
 from highdicom.pixels import apply_lut
 from highdicom.sr.coding import CodedConcept
 from highdicom.sr.value_types import CodeContentItem
-from highdicom.uid import UID
-from highdicom.utils import compute_plane_position_slide_per_frame
+from highdicom.image import DimensionIndexSequence as BaseDimensionIndexSequence
 from highdicom.valuerep import (
     _check_long_string,
     _check_short_string,
@@ -243,7 +239,7 @@ class RealWorldValueMapping(Dataset):
             return array * slope + intercept
 
 
-class DimensionIndexSequence(DataElementSequence):
+class DimensionIndexSequence(BaseDimensionIndexSequence):
 
     """Sequence of data elements describing dimension indices for the patient
     or slide coordinate system based on the Dimension Index functional
@@ -253,273 +249,41 @@ class DimensionIndexSequence(DataElementSequence):
     ----
     The order of indices is fixed.
 
+    Note
+    ----
+    This class is deprecated and will be removed in a future version of
+    highdicom. User code should generally avoid this class, and if necessary,
+    the more general :class:`highdicom.DimensionIndexSequence` should be used
+    instead.
+
     """
 
     def __init__(
         self,
-        coordinate_system: str | CoordinateSystemNames
+        coordinate_system: str | CoordinateSystemNames | None,
     ) -> None:
         """
         Parameters
         ----------
-        coordinate_system: Union[str, highdicom.CoordinateSystemNames]
+        coordinate_system: Union[str, highdicom.CoordinateSystemNames, None]
             Subject (``"PATIENT"`` or ``"SLIDE"``) that was the target of
-            imaging
-        """
-        super().__init__()
-        dim_uid = UID()
-
-        self._coordinate_system = CoordinateSystemNames(coordinate_system)
-        if self._coordinate_system == CoordinateSystemNames.SLIDE:
-            x_axis_index = Dataset()
-            x_axis_index.DimensionIndexPointer = tag_for_keyword(
-                'XOffsetInSlideCoordinateSystem'
-            )
-            x_axis_index.FunctionalGroupPointer = tag_for_keyword(
-                'PlanePositionSlideSequence'
-            )
-            x_axis_index.DimensionOrganizationUID = dim_uid
-            x_axis_index.DimensionDescriptionLabel = \
-                'X Offset in Slide Coordinate System'
-
-            y_axis_index = Dataset()
-            y_axis_index.DimensionIndexPointer = tag_for_keyword(
-                'YOffsetInSlideCoordinateSystem'
-            )
-            y_axis_index.FunctionalGroupPointer = tag_for_keyword(
-                'PlanePositionSlideSequence'
-            )
-            y_axis_index.DimensionOrganizationUID = dim_uid
-            y_axis_index.DimensionDescriptionLabel = \
-                'Y Offset in Slide Coordinate System'
-
-            z_axis_index = Dataset()
-            z_axis_index.DimensionIndexPointer = tag_for_keyword(
-                'ZOffsetInSlideCoordinateSystem'
-            )
-            z_axis_index.FunctionalGroupPointer = tag_for_keyword(
-                'PlanePositionSlideSequence'
-            )
-            z_axis_index.DimensionOrganizationUID = dim_uid
-            z_axis_index.DimensionDescriptionLabel = \
-                'Z Offset in Slide Coordinate System'
-
-            row_dimension_index = Dataset()
-            row_dimension_index.DimensionIndexPointer = tag_for_keyword(
-                'ColumnPositionInTotalImagePixelMatrix'
-            )
-            row_dimension_index.FunctionalGroupPointer = tag_for_keyword(
-                'PlanePositionSlideSequence'
-            )
-            row_dimension_index.DimensionOrganizationUID = dim_uid
-            row_dimension_index.DimensionDescriptionLabel = \
-                'Column Position In Total Image Pixel Matrix'
-
-            column_dimension_index = Dataset()
-            column_dimension_index.DimensionIndexPointer = tag_for_keyword(
-                'RowPositionInTotalImagePixelMatrix'
-            )
-            column_dimension_index.FunctionalGroupPointer = tag_for_keyword(
-                'PlanePositionSlideSequence'
-            )
-            column_dimension_index.DimensionOrganizationUID = dim_uid
-            column_dimension_index.DimensionDescriptionLabel = \
-                'Row Position In Total Image Pixel Matrix'
-
-            # Organize frames for each segment similar to TILED_FULL, first
-            # along the row dimension (column indices from left to right) and
-            # then along the column dimension (row indices from top to bottom)
-            # of the Total Pixel Matrix.
-            self.extend([
-                row_dimension_index,
-                column_dimension_index,
-                x_axis_index,
-                y_axis_index,
-                z_axis_index,
-            ])
-
-        elif self._coordinate_system == CoordinateSystemNames.PATIENT:
-            image_position_index = Dataset()
-            image_position_index.DimensionIndexPointer = tag_for_keyword(
-                'ImagePositionPatient'
-            )
-            image_position_index.FunctionalGroupPointer = tag_for_keyword(
-                'PlanePositionSequence'
-            )
-            image_position_index.DimensionOrganizationUID = dim_uid
-            image_position_index.DimensionDescriptionLabel = \
-                'Image Position Patient'
-
-            self.append(image_position_index)
-
-        else:
-            raise ValueError(
-                f'Unknown coordinate system "{self._coordinate_system}"'
-            )
-
-    def get_plane_positions_of_image(
-        self,
-        image: Dataset
-    ) -> list[PlanePositionSequence]:
-        """Get plane positions of frames in multi-frame image.
-
-        Parameters
-        ----------
-        image: Dataset
-            Multi-frame image
-
-        Returns
-        -------
-        List[highdicom.PlanePositionSequence]
-            Plane position of each frame in the image
+            imaging. If None, the imaging does not belong within a frame of
+            reference.
+        include_segment_number: bool
+            Include the segment number as a dimension index.
 
         """
-        is_multiframe = is_multiframe_image(image)
-        if not is_multiframe:
-            raise ValueError('Argument "image" must be a multi-frame image.')
-
-        if self._coordinate_system == CoordinateSystemNames.SLIDE:
-            if hasattr(image, 'PerFrameFunctionalGroupsSequence'):
-                plane_positions = [
-                    item.PlanePositionSlideSequence
-                    for item in image.PerFrameFunctionalGroupsSequence
-                ]
-            else:
-                # If Dimension Organization Type is TILED_FULL, plane
-                # positions are implicit and need to be computed.
-                plane_positions = compute_plane_position_slide_per_frame(
-                    image
-                )
-        else:
-            plane_positions = [
-                item.PlanePositionSequence
-                for item in image.PerFrameFunctionalGroupsSequence
-            ]
-
-        return plane_positions
-
-    def get_plane_positions_of_series(
-        self,
-        images: Sequence[Dataset]
-    ) -> list[PlanePositionSequence]:
-        """Gets plane positions for series of single-frame images.
-
-        Parameters
-        ----------
-        images: Sequence[Dataset]
-            Series of single-frame images
-
-        Returns
-        -------
-        List[highdicom.PlanePositionSequence]
-            Plane position of each frame in the image
-
-        """
-        is_multiframe = any([is_multiframe_image(img) for img in images])
-        if is_multiframe:
-            raise ValueError(
-                'Argument "images" must be a series of single-frame images.'
-            )
-
-        plane_positions = [
-            PlanePositionSequence(
-                coordinate_system=CoordinateSystemNames.PATIENT,
-                image_position=img.ImagePositionPatient
-            )
-            for img in images
-        ]
-
-        return plane_positions
-
-    def get_index_values(
-        self,
-        plane_positions: Sequence[PlanePositionSequence]
-    ) -> tuple[np.ndarray, np.ndarray]:
-        """Get the values of indexed attributes.
-
-        Parameters
-        ----------
-        plane_positions: Sequence[highdicom.PlanePositionSequence]
-            Plane position of frames in a multi-frame image or in a series of
-            single-frame images
-
-        Returns
-        -------
-        dimension_index_values: numpy.ndarray
-            2D array of spatial dimension index values
-        plane_indices: numpy.ndarray
-            1D array of planes indices for sorting frames according to their
-            spatial position specified by the dimension index.
-
-        """
-        # For each indexed spatial dimension  obtain the value of the attribute
-        # that the Dimension Index Pointer points to in the element of the
-        # Plane Position Sequence or Plane Position Slide Sequence.
-        # In case of the patient coordinate system, this is the Image Position
-        # Patient attribute. In case of the slide coordinate system, these are
-        # X/Y/Z Offset In Slide Coordinate System and the Column/Row
-        # Position in Total Image Pixel Matrix attributes.
-        plane_position_values = np.array([
-            [
-                np.array(p[0][indexer.DimensionIndexPointer].value)
-                for indexer in self
-            ]
-            for p in plane_positions
-        ])
-
-        # Build an array that can be used to sort planes according to the
-        # Dimension Index Value based on the order of the items in the
-        # Dimension Index Sequence.
-        _, plane_sort_indices = np.unique(
-            plane_position_values,
-            axis=0,
-            return_index=True
+        warnings.warn(
+            "The highdicom.pm.DimensionIndexSequence class is deprecated and "
+            "will be removed in a future version of the library. User code "
+            "should typically avoid this class, or, if required, use the more "
+            "general highdicom.DimensionIndexSequence instead.",
+            UserWarning,
+            stacklevel=2,
         )
-
-        return (plane_position_values, plane_sort_indices)
-
-    def get_index_position(self, pointer: str) -> int:
-        """Get relative position of a given dimension in the dimension index.
-
-        Parameters
-        ----------
-        pointer: str
-            Name of the dimension (keyword of the attribute),
-            e.g., ``"XOffsetInSlideCoordinateSystem"``
-
-        Returns
-        -------
-        int
-            Zero-based relative position
-
-        Examples
-        --------
-        >>> dimension_index = DimensionIndexSequence("SLIDE")
-        >>> i = dimension_index.get_index_position("XOffsetInSlideCoordinateSystem")
-        >>> x_offsets = dimension_index[i]
-
-        """  # noqa: E501
-        indices = [
-            i
-            for i, indexer in enumerate(self)
-            if indexer.DimensionIndexPointer == tag_for_keyword(pointer)
-        ]
-        if len(indices) == 0:
-            raise ValueError(
-                f'Dimension index does not contain a dimension "{pointer}".'
-            )
-        return indices[0]
-
-    def get_index_keywords(self) -> list[str]:
-        """Get keywords of attributes that specify the position of planes.
-
-        Returns
-        -------
-        List[str]
-            Keywords of indexed attributes
-
-        """
-        return [
-            keyword_for_tag(indexer.DimensionIndexPointer)
-            for indexer in self
-        ]
+        super().__init__(
+            coordinate_system=coordinate_system,
+            functional_groups_module=(
+                'segmentation-multi-frame-functional-groups'
+            ),
+        )
